Refactor CommentContents
