+++ src/components/Comment/CommentContent/CommentContent.module.scss
@@ -35,3 +35,10 @@
     }
   }
 }
+
+.Reply {
+  .ContentContainer {
+    margin-left: 2rem;
+    background-color: var(--secondary);
+  }
+}
