+++ src/components/Comment/CommentContent/index.tsx
@@ -1,9 +1,9 @@
 import { useState } from 'react';
 import { ICommentContent } from '../index';
-import Reply from '../Reply';
+import Reply from './Reply';
 import CommentInfo from './CommentInfo';
 import CommentInputForm from './CommentInputForm';
-import styles from './Content.module.scss';
+import styles from './CommentContent.module.scss';
 
 interface Props {
   comment: ICommentContent;
@@ -21,10 +21,11 @@ export const ContentText: React.FC<IContentText> = ({ text }) => {
   );
 };
 
-const Comment: React.FC<Props> = ({ comment }) => {
+const Content: React.FC<Props> = ({ comment }) => {
   const { info, reply } = comment;
   const { user, date } = info;
   const [showReply, setshowReply] = useState(false);
+
   const toggleReply = () => {
     setshowReply(!showReply);
   };
@@ -33,21 +34,10 @@ const Comment: React.FC<Props> = ({ comment }) => {
     <div className={styles.Content}>
       <CommentInfo user={user} date={date} />
       <ContentText text={comment.text} />
-      
-      <div className={styles.replyContainer}>
-        {reply ? (
-          <div className={styles.toggleButton} onClick={toggleReply}>
-            <span>{showReply ? '🐄  접기' : `🐮 ${reply.length} 개의 댓글`}</span>
-            {showReply && reply.map((item, index) => <Reply info={item.info} text={item.text} like={item.like} key={index} />)}
-          </div>
-        ) : (
-          <div className={styles.noreply}>
-            <CommentInputForm />
-          </div>
-        )}
-      </div>
+      {reply && <Reply reply={reply} showReply={showReply} toggleReply={toggleReply} />}
+      <CommentInputForm />
     </div>
   );
 };
 
-export default Comment;
+export default Content;
